test(check): the slot test moves to its own module, under the file cap
The rebase onto main tipped `check/tests.rs` past the per-file wall:

    12 file-loc-cap  RED: file(s) >= 1500 LOC without valid LOC-EXEMPT:
      crates/nika-cli/src/verbs/check/tests.rs: 1528 LOC

Same shape as the crate wall one level down: main's copy and this
branch's +31 are each under 1500, and together they are over. Only the
merge sees it, and hygiene vector 12 is the only thing that looks.

The SLOT test moves to `check/slot_tests.rs`. `mod repair_tests` already
sits beside `mod tests` in this very file, so a topic-named sibling is
the module's own precedent rather than a new convention. `checked_output`
becomes `pub(super)` -- the one helper the moved test needs that lives in
`tests.rs` rather than in `check/mod.rs`.

Verified: `cargo test -p nika-cli --lib unfilled_slot` -> 1 passed, and
vector 12 is YELLOW (cohesion suggestions only), no RED.

A LOC-EXEMPT marker was NOT the answer. The policy allows three reasons
-- codegen, lookup-table, enum-mega -- and a test file is none of them.
An exemption would have bought the green by widening the rule instead of
obeying it.

WHAT THIS DOES NOT FIX: tests.rs is now 1497 of 1500. Three lines. The
next test written there hits the same wall, and the honest split is a
thematic one across the whole file, not one test lifted off the top. I
moved only what this branch added, because choosing which of someone
else's tests to relocate is not a rebase's call. Named here so the
margin is visible rather than discovered.
